+++ sdk/textanalytics/azure-ai-textanalytics/src/main/java/com/azure/ai/textanalytics/TextAnalyticsAsyncClient.java
@@ -254,7 +254,7 @@ public final class TextAnalyticsAsyncClient {
      *
      * For a list of supported entity types, check: <a href="https://aka.ms/taner">this</a>.
      * For a list of enabled languages, check: <a href="https://aka.ms/talangs">this</a>.
-     * This method will use the default language that sets up in
+     * This method will use the default language that is set using
      * {@link TextAnalyticsClientBuilder#defaultLanguage(String)}. If none is specified, service will use 'en' as
      * the language.
      *
@@ -376,7 +376,7 @@ public final class TextAnalyticsAsyncClient {
      *
      * For a list of supported entity types, check: <a href="https://aka.ms/tanerpii">this</a>.
      * For a list of enabled languages, check: <a href="https://aka.ms/talangs">this</a>. This method will use the
-     * default language that sets up in {@link TextAnalyticsClientBuilder#defaultLanguage(String)}. If none is
+     * default language that is set using {@link TextAnalyticsClientBuilder#defaultLanguage(String)}. If none is
      * specified, service will use 'en' as the language.
      *
      * <p><strong>Code sample</strong></p>
@@ -500,7 +500,7 @@ public final class TextAnalyticsAsyncClient {
      * Returns a list of recognized entities with links to a well-known knowledge base for the provided document. See
      * <a href="https://aka.ms/talangs">this</a> for supported languages in Text Analytics API.
      *
-     * This method will use the default language that sets up in
+     * This method will use the default language that is set using
      * {@link TextAnalyticsClientBuilder#defaultLanguage(String)}. If none is specified, service will use 'en' as
      * the language.
      *
@@ -620,7 +620,7 @@ public final class TextAnalyticsAsyncClient {
     /**
      * Returns a list of strings denoting the key phrases in the document.
      *
-     * This method will use the default language that sets up in
+     * This method will use the default language that is set using
      * {@link TextAnalyticsClientBuilder#defaultLanguage(String)}. If none is specified, service will use 'en' as
      * the language.
      *
@@ -739,7 +739,7 @@ public final class TextAnalyticsAsyncClient {
      * Returns a sentiment prediction, as well as confidence scores for each sentiment label (Positive, Negative, and
      * Neutral) for the document and each sentence within it.
      *
-     * This method will use the default language that sets up in
+     * This method will use the default language that is set using
      * {@link TextAnalyticsClientBuilder#defaultLanguage(String)}. If none is specified, service will use 'en' as
      * the language.
      *