+++ sdk/textanalytics/azure-ai-textanalytics/src/main/java/com/azure/ai/textanalytics/models/PiiEntity.java
@@ -26,6 +26,11 @@ public final class PiiEntity {
     private final String subcategory;
 
     /*
+     * Confidence score between 0 and 1 of the extracted entity.
+     */
+    private final double confidenceScore;
+
+    /*
      * Start position for the entity text.
      */
     private final int offset;
@@ -35,23 +40,18 @@ public final class PiiEntity {
      */
     private final int length;
 
-    /*
-     * Confidence score between 0 and 1 of the extracted entity.
-     */
-    private final double confidenceScore;
-
     /**
      * Creates a {@link PiiEntity} model that describes entity.
      *
      * @param text The entity text as appears in the request.
      * @param category The entity category, such as Person/Location/Org/SSN etc.
      * @param subcategory The entity subcategory, such as Medical/Stock exchange/Sports etc.
+     * @param confidenceScore A confidence score between 0 and 1 of the recognized entity.
      * @param offset The start position for the entity text
      * @param length The length for the entity text
-     * @param confidenceScore A confidence score between 0 and 1 of the extracted entity.
      */
-    public PiiEntity(String text, EntityCategory category, String subcategory, int offset, int length,
-        double confidenceScore) {
+    public PiiEntity(String text, EntityCategory category, String subcategory, double confidenceScore, int offset,
+        int length) {
         this.text = text;
         this.category = category;
         this.subcategory = subcategory;
@@ -63,7 +63,7 @@ public final class PiiEntity {
     /**
      * Get the text property: PII entity text as appears in the request.
      *
-     * @return The text value.
+     * @return The {@code text} value.
      */
     public String getText() {
         return this.text;
@@ -72,25 +72,34 @@ public final class PiiEntity {
     /**
      * Get the category property: Categorized entity category, such as Person/Location/Org/SSN etc.
      *
-     * @return The category value.
+     * @return The {@code category} value.
      */
     public EntityCategory getCategory() {
         return this.category;
     }
 
     /**
-     * Get the subcategory property: Categorized entity sub category, such as Medical/Stock exchange/Sports etc.
+     * Get the subcategory property: Categorized entity subcategory, such as Medical/Stock exchange/Sports etc.
      *
-     * @return The subcategory value.
+     * @return The {@code subcategory} value.
      */
     public String getSubcategory() {
         return this.subcategory;
     }
 
     /**
+     * Get the score property: Confidence score between 0 and 1 of the recognized entity.
+     *
+     * @return The {@code confidenceScore} value.
+     */
+    public double getConfidenceScore() {
+        return this.confidenceScore;
+    }
+
+    /**
      * Get the offset property: the start position for the entity text.
      *
-     * @return The offset value.
+     * @return The {@code offset} value.
      */
     public int getOffset() {
         return this.offset;
@@ -99,18 +108,9 @@ public final class PiiEntity {
     /**
      * Get the length property: the length for the entity text.
      *
-     * @return The length value.
+     * @return The {@code length} value.
      */
     public int getLength() {
         return this.length;
-    }
-
-    /**
-     * Get the score property: Confidence score between 0 and 1 of the extracted entity.
-     *
-     * @return The score value.
-     */
-    public double getConfidenceScore() {
-        return this.confidenceScore;
     }
 }