+++ sdk/textanalytics/azure-ai-textanalytics/src/samples/java/com/azure/ai/textanalytics/ReadmeSamples.java
@@ -155,9 +155,10 @@ public class ReadmeSamples {
      * Code snippet for recognizing Personally Identifiable Information entity in a document.
      */
     public void recognizePiiEntity() {
-        String document = "My SSN is 555-55-5555";
+        String document = "My SSN is 859-98-0987";
         textAnalyticsClient.recognizePiiEntities(document).forEach(entity -> System.out.printf(
-            "Recognized Personally Identifiable Information entity: %s, entity category: %s, entity subcategory: %s, offset: %s, length: %s, confidence score: %f.%n",
-            entity.getText(), entity.getCategory(), entity.getSubcategory(), entity.getOffset(), entity.getLength(), entity.getConfidenceScore()));
+            "Recognized Personally Identifiable Information entity: %s, entity category: %s, entity subcategory: %s,"
+                + " confidence score: %f.%n",
+            entity.getText(), entity.getCategory(), entity.getSubcategory(), entity.getConfidenceScore()));
     }
 }