+++ sdk/textanalytics/azure-ai-textanalytics/src/samples/java/com/azure/ai/textanalytics/RecognizePiiEntities.java
@@ -24,10 +24,10 @@ public class RecognizePiiEntities {
             .buildClient();
 
         // The document that needs be analyzed.
-        String document = "My SSN is 555-55-5555";
+        String document = "My SSN is 859-98-0987";
 
         client.recognizePiiEntities(document).forEach(entity -> System.out.printf(
-            "Recognized personal identifiable information entity: %s, entity category: %s, entity sub-category: %s, score: %f.%n",
+            "Recognized Personal Identifiable Information entity: %s, entity category: %s, entity sub-category: %s, score: %f.%n",
             entity.getText(), entity.getCategory(), entity.getSubcategory(), entity.getConfidenceScore()));
     }
 }